FIPS: Use nop keystore implementation
Keystore.Factory and other constuctor functions will return a
NopKeystore if the requirefips build tag is set. The NopKeystore will
will return ErrNotWritable for any retrievals, and
ErrKeystoreUnsupported for create/save/store operations. The NopKeystore
uses the same existing constructor name as the regular file keystore so
it can be used in-place by callers.
