+++ keystore/file_keystore_nofips_test.go
